llm/requirements.txt
requests>=2.31.0
gradio>=4.0.0
pyyaml>=6.0
